# editoast: infra: order objects when downloading a railjson